Remove OpenRouter Models by Provider backup
