+++ packages/db/src/schema.ts
@@ -1994,6 +1994,8 @@ export const modelStats = pgTable(
 export type ModelStats = typeof modelStats.$inferSelect;
 export type NewModelStats = typeof modelStats.$inferInsert;
 
+export const MODELS_BY_PROVIDER_SCRIPT_NAME = 'pnpm script:run openrouter sync-providers';
+
 export const modelsByProvider = pgTable('models_by_provider', {
   id: serial().notNull().primaryKey(),
   data: jsonb('data').$type<NormalizedOpenRouterResponse>().notNull(),
